feat: add global success dialog box
