+++ .env
@@ -1,3 +1,2 @@
-#VITE_APP_BASE_URL=http://localhost
+VITE_APP_BASE_URL=http://localhost
 VITE_APP_BASE_PORT=80
-VITE_APP_BASE_URL=http://192.168.0.20
