+++ src/components/Profile/Dialogs/index.ts
@@ -2,4 +2,3 @@ export * from './Avatar';
 export * from './Dialog';
 export * from './Name';
 export * from './Password';
-export * from './Success';
