+++ src/layouts/index.ts
@@ -4,3 +4,4 @@ export { Sidebar } from './Sidebar';
 export { Auth } from './Auth';
 export { Loading } from '@app/layouts/Loading';
 export { Error } from '@app/layouts/Error';
+export { Success } from '@app/layouts/Success';
